+++ src/intrpr/cfg_mgr.py
@@ -56,7 +56,7 @@ def get_cfg() -> Cfg:
 
     cfg = sandbox_runpy(uconst.CFG_FL)
     if cfg is None:
-        return Cfg(prompt=prompt, pth=pth)
+        return Cfg(prompt=prompt, pth=pth, aliases=aliases)
 
     for key in cfg:
         val = cfg[key]